Move code validation into modules
