--- pxtblocks/code-validation/evaluationResult.ts
@@ -1,5 +1,0 @@
-namespace pxt.blocks {
-  export interface EvaluationResult {
-   result: boolean;
-  }
-}
